# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for ITSM",
        "title": "ITSM API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/v1/api/btcusdt": {
            "get": {
                "description": "Получить последний актуальный курс BTS_USDT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Валюты"
                ],
                "summary": "Последняя запись",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BTCUSDTResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Получить историю изменений поддерживает фильтрацию по и дате и времени\nввести дату и время в поле Date в формате yyy-mm-dd hh:mm:ss",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Валюты"
                ],
                "summary": "история Изменений",
                "parameters": [
                    {
                        "description": "info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.BTCUSDBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.BTCUSDTResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/api/currencies": {
            "get": {
                "description": "Получить актуальные крусы BTC к фиатным валютам",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Валюты"
                ],
                "summary": "Последняя запись",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.CurrencyDec"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Получить историю изменений поддерживает фильтрацию по валютам и дате\nчтобы использовать нужно указать через запятую в двойных кавычках и заглавными буквами\nназвания валют в поле Name, а так же дату в поле Date в формате yyy-mm-dd",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Валюты"
                ],
                "summary": "история Изменений",
                "parameters": [
                    {
                        "description": "info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CurrencyBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.CurrencyResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/api/latest": {
            "get": {
                "description": "Получить последние актиальные курсы валют",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Валюты"
                ],
                "summary": "Последняя запись",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.CurrencyDec"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.BTCUSDBody": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "entity.BTCUSDTResponse": {
            "type": "object",
            "properties": {
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.BTCUSDTResult"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "entity.BTCUSDTResult": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "entity.CurrencyBody": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "name": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "entity.CurrencyDec": {
            "type": "object",
            "properties": {
                "AMD": {
                    "type": "number"
                },
                "AUD": {
                    "type": "number"
                },
                "AZN": {
                    "type": "number"
                },
                "BGN": {
                    "type": "number"
                },
                "BRL": {
                    "type": "number"
                },
                "BYN": {
                    "type": "number"
                },
                "CAD": {
                    "type": "number"
                },
                "CHF": {
                    "type": "number"
                },
                "CNY": {
                    "type": "number"
                },
                "CZK": {
                    "type": "number"
                },
                "DKK": {
                    "type": "number"
                },
                "Date": {
                    "type": "string"
                },
                "EUR": {
                    "type": "number"
                },
                "GBP": {
                    "type": "number"
                },
                "HKD": {
                    "type": "number"
                },
                "HUF": {
                    "type": "number"
                },
                "INR": {
                    "type": "number"
                },
                "JPY": {
                    "type": "number"
                },
                "KGS": {
                    "type": "number"
                },
                "KRW": {
                    "type": "number"
                },
                "KZT": {
                    "type": "number"
                },
                "MDL": {
                    "type": "number"
                },
                "NOK": {
                    "type": "number"
                },
                "PLN": {
                    "type": "number"
                },
                "RON": {
                    "type": "number"
                },
                "SEK": {
                    "type": "number"
                },
                "SGD": {
                    "type": "number"
                },
                "TJS": {
                    "type": "number"
                },
                "TMT": {
                    "type": "number"
                },
                "TRY": {
                    "type": "number"
                },
                "UAH": {
                    "type": "number"
                },
                "USD": {
                    "type": "number"
                },
                "UZS": {
                    "type": "number"
                },
                "XDR": {
                    "type": "number"
                },
                "ZAR": {
                    "type": "number"
                }
            }
        },
        "entity.CurrencyResponse": {
            "type": "object",
            "properties": {
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CurrencyDec"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}